Seperate logic for fetching movies and weath^C forecasts into their own modules.
